feat: abstract helper functions into importanceUtils.js; update exports for the hooks and utils folders; add comments to setImportantList
