+++ src/utils/index.js
@@ -1,2 +1,3 @@
 export * from './dates';
 export * from './normalize';
+export * from './importanceUtils';
